requirements.txt
opencv-python
mediapipe
numpy
langchain
langchain-openai
python-dotenv
streamlit
